+++ hs-bindgen/test-artefacts/fixtures/macros/reparse/nesting/union_in_typedef.2.raw/Example.hs
@@ -78,7 +78,9 @@ instance (ty ~ BG.CInt) => BG.HasField "t1_x" T1 ty where
 instance (ty ~ BG.CInt) => BG.CompatHasField.HasField "t1_x" T1 ty where
 
   hasField =
-    \x0 -> (BG.setUnionPayload, BG.getField @"t1_x" x0)
+    \x0 ->
+      (\y1 ->
+         BG.setUnionPayload y1 x0, BG.getField @"t1_x" x0)
 
 instance (ty ~ BG.CInt) => BG.HasField "t1_x" (BG.Ptr T1) (BG.Ptr ty) where
 
@@ -131,7 +133,8 @@ instance (ty ~ BG.CInt) => BG.CompatHasField.HasField "t2_Aux_x" T2_Aux ty where
 
   hasField =
     \x0 ->
-      (BG.setUnionPayload, BG.getField @"t2_Aux_x" x0)
+      (\y1 ->
+         BG.setUnionPayload y1 x0, BG.getField @"t2_Aux_x" x0)
 
 instance ( ty ~ BG.CInt
          ) => BG.HasField "t2_Aux_x" (BG.Ptr T2_Aux) (BG.Ptr ty) where
@@ -222,7 +225,8 @@ instance (ty ~ BG.CInt) => BG.CompatHasField.HasField "t3_Aux_x" T3_Aux ty where
 
   hasField =
     \x0 ->
-      (BG.setUnionPayload, BG.getField @"t3_Aux_x" x0)
+      (\y1 ->
+         BG.setUnionPayload y1 x0, BG.getField @"t3_Aux_x" x0)
 
 instance ( ty ~ BG.CInt
          ) => BG.HasField "t3_Aux_x" (BG.Ptr T3_Aux) (BG.Ptr ty) where