+++ hs-bindgen/test-artefacts/fixtures/macros/reparse/nesting/union_in_union.1.empty/Example.hs
@@ -68,7 +68,9 @@ instance (ty ~ BG.CInt) => BG.HasField "t1_x_x" T1_x ty where
 instance (ty ~ BG.CInt) => BG.CompatHasField.HasField "t1_x_x" T1_x ty where
 
   hasField =
-    \x0 -> (BG.setUnionPayload, BG.getField @"t1_x_x" x0)
+    \x0 ->
+      (\y1 ->
+         BG.setUnionPayload y1 x0, BG.getField @"t1_x_x" x0)
 
 instance (ty ~ BG.CInt) => BG.HasField "t1_x_x" (BG.Ptr T1_x) (BG.Ptr ty) where
 
@@ -120,7 +122,9 @@ instance (ty ~ T1_x) => BG.HasField "t1_x" T1 ty where
 instance (ty ~ T1_x) => BG.CompatHasField.HasField "t1_x" T1 ty where
 
   hasField =
-    \x0 -> (BG.setUnionPayload, BG.getField @"t1_x" x0)
+    \x0 ->
+      (\y1 ->
+         BG.setUnionPayload y1 x0, BG.getField @"t1_x" x0)
 
 instance (ty ~ T1_x) => BG.HasField "t1_x" (BG.Ptr T1) (BG.Ptr ty) where
 
@@ -172,7 +176,9 @@ instance (ty ~ BG.Ptr T2_x) => BG.HasField "t2_x" T2 ty where
 instance (ty ~ BG.Ptr T2_x) => BG.CompatHasField.HasField "t2_x" T2 ty where
 
   hasField =
-    \x0 -> (BG.setUnionPayload, BG.getField @"t2_x" x0)
+    \x0 ->
+      (\y1 ->
+         BG.setUnionPayload y1 x0, BG.getField @"t2_x" x0)
 
 instance (ty ~ BG.Ptr T2_x) => BG.HasField "t2_x" (BG.Ptr T2) (BG.Ptr ty) where
 
@@ -224,7 +230,9 @@ instance (ty ~ BG.CInt) => BG.HasField "t2_x_x" T2_x ty where
 instance (ty ~ BG.CInt) => BG.CompatHasField.HasField "t2_x_x" T2_x ty where
 
   hasField =
-    \x0 -> (BG.setUnionPayload, BG.getField @"t2_x_x" x0)
+    \x0 ->
+      (\y1 ->
+         BG.setUnionPayload y1 x0, BG.getField @"t2_x_x" x0)
 
 instance (ty ~ BG.CInt) => BG.HasField "t2_x_x" (BG.Ptr T2_x) (BG.Ptr ty) where
 
@@ -277,7 +285,9 @@ instance ( ty ~ BG.Ptr (BG.Ptr T3_x)
          ) => BG.CompatHasField.HasField "t3_x" T3 ty where
 
   hasField =
-    \x0 -> (BG.setUnionPayload, BG.getField @"t3_x" x0)
+    \x0 ->
+      (\y1 ->
+         BG.setUnionPayload y1 x0, BG.getField @"t3_x" x0)
 
 instance ( ty ~ BG.Ptr (BG.Ptr T3_x)
          ) => BG.HasField "t3_x" (BG.Ptr T3) (BG.Ptr ty) where
@@ -330,7 +340,9 @@ instance (ty ~ BG.CInt) => BG.HasField "t3_x_x" T3_x ty where
 instance (ty ~ BG.CInt) => BG.CompatHasField.HasField "t3_x_x" T3_x ty where
 
   hasField =
-    \x0 -> (BG.setUnionPayload, BG.getField @"t3_x_x" x0)
+    \x0 ->
+      (\y1 ->
+         BG.setUnionPayload y1 x0, BG.getField @"t3_x_x" x0)
 
 instance (ty ~ BG.CInt) => BG.HasField "t3_x_x" (BG.Ptr T3_x) (BG.Ptr ty) where