+++ hs-bindgen/test-artefacts/fixtures/macros/reparse/nesting/union_in_union.2.raw/Example.hs
@@ -78,7 +78,9 @@ instance (ty ~ BG.CInt) => BG.HasField "t1_x_x" T1_x ty where
 instance (ty ~ BG.CInt) => BG.CompatHasField.HasField "t1_x_x" T1_x ty where
 
   hasField =
-    \x0 -> (BG.setUnionPayload, BG.getField @"t1_x_x" x0)
+    \x0 ->
+      (\y1 ->
+         BG.setUnionPayload y1 x0, BG.getField @"t1_x_x" x0)
 
 instance (ty ~ BG.CInt) => BG.HasField "t1_x_x" (BG.Ptr T1_x) (BG.Ptr ty) where
 
@@ -130,7 +132,9 @@ instance (ty ~ T1_x) => BG.HasField "t1_x" T1 ty where
 instance (ty ~ T1_x) => BG.CompatHasField.HasField "t1_x" T1 ty where
 
   hasField =
-    \x0 -> (BG.setUnionPayload, BG.getField @"t1_x" x0)
+    \x0 ->
+      (\y1 ->
+         BG.setUnionPayload y1 x0, BG.getField @"t1_x" x0)
 
 instance (ty ~ T1_x) => BG.HasField "t1_x" (BG.Ptr T1) (BG.Ptr ty) where
 
@@ -182,7 +186,9 @@ instance (ty ~ BG.Ptr T2_x) => BG.HasField "t2_x" T2 ty where
 instance (ty ~ BG.Ptr T2_x) => BG.CompatHasField.HasField "t2_x" T2 ty where
 
   hasField =
-    \x0 -> (BG.setUnionPayload, BG.getField @"t2_x" x0)
+    \x0 ->
+      (\y1 ->
+         BG.setUnionPayload y1 x0, BG.getField @"t2_x" x0)
 
 instance (ty ~ BG.Ptr T2_x) => BG.HasField "t2_x" (BG.Ptr T2) (BG.Ptr ty) where
 
@@ -234,7 +240,9 @@ instance (ty ~ BG.CInt) => BG.HasField "t2_x_x" T2_x ty where
 instance (ty ~ BG.CInt) => BG.CompatHasField.HasField "t2_x_x" T2_x ty where
 
   hasField =
-    \x0 -> (BG.setUnionPayload, BG.getField @"t2_x_x" x0)
+    \x0 ->
+      (\y1 ->
+         BG.setUnionPayload y1 x0, BG.getField @"t2_x_x" x0)
 
 instance (ty ~ BG.CInt) => BG.HasField "t2_x_x" (BG.Ptr T2_x) (BG.Ptr ty) where
 
@@ -287,7 +295,9 @@ instance ( ty ~ BG.Ptr (BG.Ptr T3_x)
          ) => BG.CompatHasField.HasField "t3_x" T3 ty where
 
   hasField =
-    \x0 -> (BG.setUnionPayload, BG.getField @"t3_x" x0)
+    \x0 ->
+      (\y1 ->
+         BG.setUnionPayload y1 x0, BG.getField @"t3_x" x0)
 
 instance ( ty ~ BG.Ptr (BG.Ptr T3_x)
          ) => BG.HasField "t3_x" (BG.Ptr T3) (BG.Ptr ty) where
@@ -340,7 +350,9 @@ instance (ty ~ BG.CInt) => BG.HasField "t3_x_x" T3_x ty where
 instance (ty ~ BG.CInt) => BG.CompatHasField.HasField "t3_x_x" T3_x ty where
 
   hasField =
-    \x0 -> (BG.setUnionPayload, BG.getField @"t3_x_x" x0)
+    \x0 ->
+      (\y1 ->
+         BG.setUnionPayload y1 x0, BG.getField @"t3_x_x" x0)
 
 instance (ty ~ BG.CInt) => BG.HasField "t3_x_x" (BG.Ptr T3_x) (BG.Ptr ty) where