+++ hs-bindgen/test-artefacts/fixtures/types/anonymous/struct_in_union/Example.hs
@@ -185,7 +185,8 @@ instance ( ty ~ BG.CChar
 
   hasField =
     \x0 ->
-      (BG.setUnionPayload, BG.getField @"outer1_fieldA" x0)
+      (\y1 ->
+         BG.setUnionPayload y1 x0, BG.getField @"outer1_fieldA" x0)
 
 instance ( ty ~ BG.CChar
          ) => BG.HasField "outer1_fieldA" (BG.Ptr Outer1) (BG.Ptr ty) where
@@ -221,7 +222,8 @@ instance ( ty ~ Outer1_anon'fieldX
 
   hasField =
     \x0 ->
-      (BG.setUnionPayload, BG.getField @"outer1_anon'fieldX" x0)
+      (\y1 ->
+         BG.setUnionPayload y1 x0, BG.getField @"outer1_anon'fieldX" x0)
 
 instance ( ty ~ Outer1_anon'fieldX
          ) => BG.HasField "outer1_anon'fieldX" (BG.Ptr Outer1) (BG.Ptr ty) where
@@ -339,7 +341,8 @@ instance ( ty ~ BG.CInt
 
   hasField =
     \x0 ->
-      (BG.setUnionPayload, BG.getField @"outer1_fieldC" x0)
+      (\y1 ->
+         BG.setUnionPayload y1 x0, BG.getField @"outer1_fieldC" x0)
 
 instance ( ty ~ BG.CInt
          ) => BG.HasField "outer1_fieldC" (BG.Ptr Outer1) (BG.Ptr ty) where
@@ -509,7 +512,8 @@ instance ( ty ~ BG.CChar
 
   hasField =
     \x0 ->
-      (BG.setUnionPayload, BG.getField @"outer2_fieldA" x0)
+      (\y1 ->
+         BG.setUnionPayload y1 x0, BG.getField @"outer2_fieldA" x0)
 
 instance ( ty ~ BG.CChar
          ) => BG.HasField "outer2_fieldA" (BG.Ptr Outer2) (BG.Ptr ty) where
@@ -544,7 +548,8 @@ instance ( ty ~ Outer2_fieldB
 
   hasField =
     \x0 ->
-      (BG.setUnionPayload, BG.getField @"outer2_fieldB" x0)
+      (\y1 ->
+         BG.setUnionPayload y1 x0, BG.getField @"outer2_fieldB" x0)
 
 instance ( ty ~ Outer2_fieldB
          ) => BG.HasField "outer2_fieldB" (BG.Ptr Outer2) (BG.Ptr ty) where
@@ -580,7 +585,8 @@ instance ( ty ~ BG.CInt
 
   hasField =
     \x0 ->
-      (BG.setUnionPayload, BG.getField @"outer2_fieldC" x0)
+      (\y1 ->
+         BG.setUnionPayload y1 x0, BG.getField @"outer2_fieldC" x0)
 
 instance ( ty ~ BG.CInt
          ) => BG.HasField "outer2_fieldC" (BG.Ptr Outer2) (BG.Ptr ty) where
@@ -744,7 +750,8 @@ instance ( ty ~ BG.CChar
 
   hasField =
     \x0 ->
-      (BG.setUnionPayload, BG.getField @"outer3_fieldA" x0)
+      (\y1 ->
+         BG.setUnionPayload y1 x0, BG.getField @"outer3_fieldA" x0)
 
 instance ( ty ~ BG.CChar
          ) => BG.HasField "outer3_fieldA" (BG.Ptr Outer3) (BG.Ptr ty) where
@@ -779,7 +786,8 @@ instance ( ty ~ Inner3
 
   hasField =
     \x0 ->
-      (BG.setUnionPayload, BG.getField @"outer3_fieldB" x0)
+      (\y1 ->
+         BG.setUnionPayload y1 x0, BG.getField @"outer3_fieldB" x0)
 
 instance ( ty ~ Inner3
          ) => BG.HasField "outer3_fieldB" (BG.Ptr Outer3) (BG.Ptr ty) where
@@ -814,7 +822,8 @@ instance ( ty ~ BG.CInt
 
   hasField =
     \x0 ->
-      (BG.setUnionPayload, BG.getField @"outer3_fieldC" x0)
+      (\y1 ->
+         BG.setUnionPayload y1 x0, BG.getField @"outer3_fieldC" x0)
 
 instance ( ty ~ BG.CInt
          ) => BG.HasField "outer3_fieldC" (BG.Ptr Outer3) (BG.Ptr ty) where