+++ hs-bindgen/test-artefacts/fixtures/types/anonymous/union/Example.hs
@@ -302,7 +302,9 @@ instance ( ty ~ BG.CInt
 
   hasField =
     \x0 ->
-      (BG.setUnionPayload, BG.getField @"has_implicit_fields_anon'x5_1_x5_1" x0)
+      ( \y1 -> BG.setUnionPayload y1 x0
+      , BG.getField @"has_implicit_fields_anon'x5_1_x5_1" x0
+      )
 
 instance ( ty ~ BG.CInt
          ) => BG.HasField "has_implicit_fields_anon'x5_1_x5_1" (BG.Ptr Has_implicit_fields_anon'x5_1) (BG.Ptr ty) where
@@ -339,7 +341,9 @@ instance ( ty ~ BG.CInt
 
   hasField =
     \x0 ->
-      (BG.setUnionPayload, BG.getField @"has_implicit_fields_anon'x5_1_x5_2" x0)
+      ( \y1 -> BG.setUnionPayload y1 x0
+      , BG.getField @"has_implicit_fields_anon'x5_1_x5_2" x0
+      )
 
 instance ( ty ~ BG.CInt
          ) => BG.HasField "has_implicit_fields_anon'x5_1_x5_2" (BG.Ptr Has_implicit_fields_anon'x5_1) (BG.Ptr ty) where
@@ -397,7 +401,8 @@ instance ( ty ~ BG.CInt
 
   hasField =
     \x0 ->
-      (BG.setUnionPayload, BG.getField @"has_implicit_fields_x1" x0)
+      (\y1 ->
+         BG.setUnionPayload y1 x0, BG.getField @"has_implicit_fields_x1" x0)
 
 instance ( ty ~ BG.CInt
          ) => BG.HasField "has_implicit_fields_x1" (BG.Ptr Has_implicit_fields) (BG.Ptr ty) where
@@ -434,7 +439,9 @@ instance ( ty ~ Has_implicit_fields_anon'x2_1
 
   hasField =
     \x0 ->
-      (BG.setUnionPayload, BG.getField @"has_implicit_fields_anon'x2_1" x0)
+      ( \y1 -> BG.setUnionPayload y1 x0
+      , BG.getField @"has_implicit_fields_anon'x2_1" x0
+      )
 
 instance ( ty ~ Has_implicit_fields_anon'x2_1
          ) => BG.HasField "has_implicit_fields_anon'x2_1" (BG.Ptr Has_implicit_fields) (BG.Ptr ty) where
@@ -557,7 +564,8 @@ instance ( ty ~ BG.CInt
 
   hasField =
     \x0 ->
-      (BG.setUnionPayload, BG.getField @"has_implicit_fields_x3" x0)
+      (\y1 ->
+         BG.setUnionPayload y1 x0, BG.getField @"has_implicit_fields_x3" x0)
 
 instance ( ty ~ BG.CInt
          ) => BG.HasField "has_implicit_fields_x3" (BG.Ptr Has_implicit_fields) (BG.Ptr ty) where
@@ -594,7 +602,9 @@ instance ( ty ~ Has_implicit_fields_anon'x4_1
 
   hasField =
     \x0 ->
-      (BG.setUnionPayload, BG.getField @"has_implicit_fields_anon'x4_1" x0)
+      ( \y1 -> BG.setUnionPayload y1 x0
+      , BG.getField @"has_implicit_fields_anon'x4_1" x0
+      )
 
 instance ( ty ~ Has_implicit_fields_anon'x4_1
          ) => BG.HasField "has_implicit_fields_anon'x4_1" (BG.Ptr Has_implicit_fields) (BG.Ptr ty) where
@@ -717,7 +727,9 @@ instance ( ty ~ Has_implicit_fields_anon'x5_1
 
   hasField =
     \x0 ->
-      (BG.setUnionPayload, BG.getField @"has_implicit_fields_anon'x5_1" x0)
+      ( \y1 -> BG.setUnionPayload y1 x0
+      , BG.getField @"has_implicit_fields_anon'x5_1" x0
+      )
 
 instance ( ty ~ Has_implicit_fields_anon'x5_1
          ) => BG.HasField "has_implicit_fields_anon'x5_1" (BG.Ptr Has_implicit_fields) (BG.Ptr ty) where
@@ -840,7 +852,8 @@ instance ( ty ~ BG.CInt
 
   hasField =
     \x0 ->
-      (BG.setUnionPayload, BG.getField @"has_implicit_fields_x5" x0)
+      (\y1 ->
+         BG.setUnionPayload y1 x0, BG.getField @"has_implicit_fields_x5" x0)
 
 instance ( ty ~ BG.CInt
          ) => BG.HasField "has_implicit_fields_x5" (BG.Ptr Has_implicit_fields) (BG.Ptr ty) where